Add support for DEIT to split_gms
It seems the DEIT course which I own has one more intro text to skip.

Not sure if its the best solution, but at least it will work for users
with that course for now. Maybe we should change it later to something
more general.

It would be good to find out if courses that were released in different
years have a different intro/outro text and whether we can detect this.

Related to https://github.com/chickendude/GlossikaNativeGLS/issues/1